# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个完整的工单管理系统API，支持用户认证、工单创建、审批、模板管理等功能。",
        "title": "Open Ticket System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tickets": {
            "get": {
                "description": "分页查询工单列表，支持筛选和排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "查询工单列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "工单名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建者",
                        "name": "creator",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "工单状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "模板ID",
                        "name": "template_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ListTicketResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "根据模板创建新的工单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "创建工单",
                "parameters": [
                    {
                        "description": "创建工单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateTicketResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "根据工单ID获取工单详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "获取工单详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TicketResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "工单不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新工单信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "更新工单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新工单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTicketResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "工单不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除工单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "删除工单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteTicketResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tickets/{id}/approval": {
            "post": {
                "description": "审批工单，支持通过、拒绝等操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "工单审批",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "审批请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "审批成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tickets/{id}/close": {
            "post": {
                "description": "关闭工单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tickets"
                ],
                "summary": "关闭工单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "工单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "关闭工单请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CloseTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CloseTicketResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ApprovalRequest": {
            "description": "工单审批请求结构",
            "type": "object",
            "required": [
                "approval_user",
                "id",
                "next_step",
                "operation"
            ],
            "properties": {
                "approval_user": {
                    "description": "审批人",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1,
                    "example": "李四"
                },
                "id": {
                    "description": "工单ID",
                    "type": "integer",
                    "example": 1
                },
                "memo": {
                    "description": "审批备注",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "同意"
                },
                "next_step": {
                    "description": "下一步骤",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "step2"
                },
                "operation": {
                    "description": "审批操作",
                    "type": "string",
                    "enum": [
                        "approve",
                        "reject"
                    ],
                    "example": "approve"
                }
            }
        },
        "models.CloseTicketRequest": {
            "type": "object",
            "required": [
                "id",
                "operator"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "memo": {
                    "type": "string",
                    "maxLength": 1000
                },
                "operator": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                }
            }
        },
        "models.CloseTicketResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.CreateTicketRequest": {
            "description": "创建工单请求结构",
            "type": "object",
            "required": [
                "creator",
                "name",
                "template_id"
            ],
            "properties": {
                "creator": {
                    "description": "创建者",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1,
                    "example": "张三"
                },
                "memo": {
                    "description": "工单备注",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "需要请假3天"
                },
                "name": {
                    "description": "工单名称",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "请假申请"
                },
                "template_id": {
                    "description": "模板ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.CreateTicketResponse": {
            "description": "创建工单响应结构",
            "type": "object",
            "properties": {
                "id": {
                    "description": "工单ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.DeleteTicketResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ListTicketResponse": {
            "description": "查询工单列表响应结构",
            "type": "object",
            "properties": {
                "list": {
                    "description": "工单列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TicketResponse"
                    }
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "models.TicketResponse": {
            "description": "工单响应结构",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2025-01-27T14:30:00Z"
                },
                "id": {
                    "description": "工单ID",
                    "type": "integer",
                    "example": 1
                },
                "memo": {
                    "description": "工单备注",
                    "type": "string",
                    "example": "需要请假3天"
                },
                "operated_users": {
                    "description": "已操作用户列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "王五"
                    ]
                },
                "operators": {
                    "description": "操作人列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "张三",
                        "李四"
                    ]
                },
                "order_num": {
                    "description": "工单号",
                    "type": "string",
                    "example": "TICKET-20250127143000"
                },
                "status": {
                    "description": "工单状态",
                    "type": "string",
                    "example": "running"
                },
                "step": {
                    "description": "当前步骤",
                    "type": "string",
                    "example": "submit"
                },
                "template_id": {
                    "description": "模板ID",
                    "type": "integer",
                    "example": 1
                },
                "uid": {
                    "description": "工单唯一标识",
                    "type": "string",
                    "example": "uid-1234567890"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2025-01-27T14:30:00Z"
                }
            }
        },
        "models.UpdateTicketRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "memo": {
                    "type": "string",
                    "maxLength": 1000
                }
            }
        },
        "models.UpdateTicketResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "请输入JWT access token，格式：Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}